+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Added support for Python 3.12
 - In setup task, added functionality to covert refplat ISO SD-WAN images to proper format so script can use them
+- In setup task, added --list parameter to list all SD-WAN software images available on CML server
 - Fixed problem where setup task might fail with "Cannot modify read-only node definition"
 - Fixed problem where add task could fail with undescriptive error "IndexError: list index out of range"
 - Updated README with installation guide for Windows